Sentinel matching is now shared
